# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: getSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a new supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierInput"
      responses:
        "201":
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "409":
          description: Conflict - duplicate phone/email/whatsapp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /suppliers/{id}:
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierInput"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict - duplicate phone/email/whatsapp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        commercialReg:
          type: string
        taxReg:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/SupplierCategory"
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - companyName
        - status
        - createdAt
        - categories
    CreateSupplierInput:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        commercialReg:
          type: string
        taxReg:
          type: string
        categoryIds:
          type: array
          items:
            type: integer
      required:
        - companyName
    UpdateSupplierInput:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        commercialReg:
          type: string
        taxReg:
          type: string
        status:
          type: string
        categoryIds:
          type: array
          items:
            type: integer
      required:
        - companyName
...
    SupplierCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success